Track current Trading View PnL tracker script
The indicator helps you track trade PnL for open positions or potential
future trades. It shows the position size, price and PnL on the chart,
as well as break even line to easily visualize where the price needs to
go before the trade can be profitable.

This script is useful for brokers that don't have a reliable integration
with Trading View where it doesn't show you PnL or currently open
positions like the case with Binance.

It can track up to three entries, showing you each entry's price, size
and PnL on the chart, and at the same time it adjusts the break even
price to account for all the entries to keep you updated on where will
the combinations of the trades will become profitable.

At the bottom right of the chart it shows you the total position size,
the average entry price and the total PnL of your symbol, to get the
overall statistics on the position.

Since Trading View doesn't allow getting live trade data from Pine
Scripts, there's a convenient label in the middle right side of the
chart to quickly go (by double clicking) to the inputs and enter your
positions, without having to go to the indicator's name and look for the
settings button.

By default, this allows you to keep track of two symbols' positions. The
total statistics will be showing both positions at the bottom right
corner as mentioned above, while only showing the buy and break even
lines only if the current chart is showing the relevant symbol.

If more positions or symbols need to be tracked, you can easily modify
the script to add more inputs for a third or a fourth symbol and the
code will be able to account for them.